CWMS-2055 - Updates to allow for a token provider with a discovered token url that the client implements (#263)
Updates to allow for a token provider with a discovered token url that the client implements
Update TokenUrlDiscoveryService to be an interface
Update TokenUrlDiscoveryService to return ApiConnectionInfo. Refactors code to pass ApiConnectionInfo for token url. This moved responsibility of creation of ssl socket data to client for discovered token url.